requirements.txt
Flask==2.1.3
selenium==4.3.0
google-cloud-translate==3.8.4
mysql-connector-python==8.0.32
sshtunnel==0.4.0
python-dotenv==1.1.0
beautifulsoup4==4.13.4
wikipedia==1.4.0
requests==2.32.3